Se saca lógica de app.js y se agregan rutas y controladores
